# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIS-CUSTOMERS",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Retrieve all customers",
                "description": "Get a list of all customers.",
                "operationId": "a29d5f0bab561196b8e3bd82422e58e6",
                "responses": {
                    "200": {
                        "description": "A list of customers"
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "description": "Create a new customer with the given details.",
                "operationId": "1d27da36718a3e9726c873633af496da",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "query",
                        "description": "Customer DNI",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 45
                        }
                    },
                    {
                        "name": "id_reg",
                        "in": "query",
                        "description": "Region ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id_com",
                        "in": "query",
                        "description": "Commune ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Customer email",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 120
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Customer name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 45
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Customer last name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 45
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Customer address",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "date_reg",
                        "in": "query",
                        "description": "Registration date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Customer status",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "A",
                                "I",
                                "trash"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "dni",
                                    "id_reg",
                                    "id_com",
                                    "email",
                                    "name",
                                    "last_name",
                                    "date_reg",
                                    "status"
                                ],
                                "properties": {
                                    "dni": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "12345678"
                                    },
                                    "id_reg": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_com": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "email": {
                                        "type": "string",
                                        "maxLength": 120,
                                        "example": "customer@example.com"
                                    },
                                    "name": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "Doe"
                                    },
                                    "address": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "123 Main St"
                                    },
                                    "date_reg": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-07"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "A",
                                            "I",
                                            "trash"
                                        ],
                                        "example": "A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer created successfully"
                    },
                    "400": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}